Base XC flying potential on thermal velocity instead of cloud cover
We need sun to get thermals. Therefore, the XC flying potential indicator was based on the cloud cover. However, we experienced that high cloud cover was degrading the value of the indicator too much.

We propose to base the XC flying potential from the thermal velocity instead of the cloud cover. The thermal velocity is computed from the boundary layer height and the heat rate, so in a sense it also contains the information about the sunshine.