Cap the boundary layer with the cloud base
In the meteograms and in the XC flying potential index, we used to overestimate the “soaring ceiling” in case of low cloud base. Basically, we were ignoring the fact that we can’t really go higher than the cloud base (in VFR).

We fix this issue by capping the boundary layer to the cloud base, in case of presence of cumulus clouds.